% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AAF}
\alias{AAF}
\title{This is data to be included in my package}
\format{An object of class \code{data.frame} with 714 rows and 20 columns.}
\usage{
AAF
}
\description{
This is data to be included in my package
}
\references{
\url{http://users.stat.umn.edu/~almquist/3811_examples/all_alaska_flights.csv}
}
\author{
Jacob Woolley \email{wooll025@umn.edu}
}
\keyword{datasets}
