% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q7.R
\name{Q7}
\alias{Q7}
\title{Three Methods of Probability Distribution}
\usage{
Q7(data, funct, interval)
}
\arguments{
\item{data}{used, type of distribution, interval}

\item{funct}{what function the user would like to use
with 0,1, and 2 corresponding with gamma cauchy and binomial}

\item{interval}{the interval across which the
optimize function will operate across}
}
\value{
Sum of distribution
}
\description{
Depending on which probability distribution method the
user wants to use, (0 for Gamma, 1 for Cauchy, 2 for Binomial)
the function takes the data and a given interval and
Returns the maximum optimized sum of the distribution
}
\details{
for gamma: Q7(x,0,c(0,10))
for Cauchy: Q7(x,1,c(-10,10))
for binomial: Q7(x,2,c(-10,10))
}
