% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools_margin.R
\name{Tools_margin}
\alias{Tools_margin}
\title{Max Min and Mean of single Column of Data}
\usage{
Tools_margin(dat, zed)
}
\arguments{
\item{zed}{column name (in quotations)}

\item{data}{Data set being used}
}
\value{
list of selected column, max value and min value in that order
}
\description{
Uses Max and min function after selecting a specific
column of data
}
